# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manhaj API",
    "version": "1.0.0",
    "description": "Offline-First Student Question Bank API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    },
    "schemas": {
...
      "Grade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Term": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gradeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/grades": {
      "post": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grade created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of grades",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Grade"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/grades/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grade updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grade deleted"
          },
          "409": {
            "description": "Cannot delete grade with dependencies"
          }
        }
      }
    },
    "/admin/terms": {
      "post": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "gradeId",
                  "name",
                  "description"
                ],
                "properties": {
                  "gradeId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Term created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gradeId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of terms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Term"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/terms/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gradeId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Term updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Term deleted"
          },
          "409": {
            "description": "Cannot delete term with dependencies"
          }
        }
      }
    },
    "/admin/modules": {
      "post": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "termId",
                  "name",
                  "description"
                ],
                "properties": {
                  "termId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "termId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of modules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Module"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/modules/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "termId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module deleted"
          },
          "409": {
            "description": "Cannot delete module with dependencies"
          }
        }
      }
    },
    "/admin/subjects": {
      "post": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "moduleId",
                  "name"
                ],
                "properties": {
                  "moduleId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subject created"
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "moduleId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of subjects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subject"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/subjects/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "moduleId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subject updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subject deleted"
          },
          "409": {
            "description": "Cannot delete subject with dependencies"
          }
        }
      }
    },
    "/admin/questions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lectureId",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions with choices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Question"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/questions/bulk-assign-lecture": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "questionIds",
                  "lectureId"
                ],
                "properties": {
                  "questionIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "lectureId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Questions assigned"
          }
        }
      }
    },
    "/admin/questions/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Question"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "authToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question deleted"
          }
        }
      }
    },
...
  },
...
}